Added mobility and tweaked gemspec to work
